+++ client/src/App.jsx
@@ -1,5 +1,4 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-
 import ChallengeList from './pages/ChallengeList';
 import UserList from './pages/UserList';
 import MyProfilePage from './pages/MyProfilePage';
@@ -9,8 +8,8 @@ import Footer from './components/Common/Footer';
 import MainPage from './pages/MainPage';
 import SearchResult from './pages/SearchResult';
 import DetailPage from './pages/DetailPage';
-import AddChallenge from './pages/AddChallenge';
 import ChallengeDetail from './pages/ChallengeDetail';
+import CreateChallengePage from './pages/CreateChallenge';
 
 function App() {
   return (
@@ -26,7 +25,7 @@ function App() {
           <Route path="/search/*" element={<SearchResult />} />
           <Route path="/profile" element={<MyProfilePage />} />
           <Route path="/ordersheet" element={<OrderSheetPage />} />
-          <Route path="/addchallenge" element={<AddChallenge />} />
+          <Route path="/createchallenge/*" element={<CreateChallengePage />} />
         </Routes>
         <Footer />
       </BrowserRouter>
