+++ client/src/components/CreateChallenge/CreateChallenge.jsx
@@ -1,10 +1,10 @@
-import { useRecoilState } from 'recoil';
+import { useRecoilValue } from 'recoil';
 import { createNumber } from '../../atoms/atoms';
 import styled from 'styled-components';
 import ChallengeAsk1 from './CreateAsk1';
 import ChallengeAsk2 from './CreateAsk2';
-import ChallengeAsk from './CreateAsk0';
 import ChallengeAsk3 from './CreateAsk3';
+import ChallengeAsk4 from './CreateAsk4';
 
 const CreateContainer = styled.section`
   display: flex;
@@ -12,14 +12,14 @@ const CreateContainer = styled.section`
 `;
 
 function CreateChallenge() {
-  const [number, setNumber] = useRecoilState(createNumber);
+  const number = useRecoilValue(createNumber);
 
   return (
     <CreateContainer>
-      {number === 0 ? <ChallengeAsk /> : null}
       {number === 1 ? <ChallengeAsk1 /> : null}
       {number === 2 ? <ChallengeAsk2 /> : null}
       {number === 3 ? <ChallengeAsk3 /> : null}
+      {number === 4 ? <ChallengeAsk4 /> : null}
     </CreateContainer>
   );
 }
